% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sncf.R
\name{get_sncf}
\alias{get_sncf}
\title{Retrieve Data From SNCF API}
\usage{
get_sncf(dataset = "ponctualite-mensuelle-transilien")
}
\arguments{
\item{dataset}{Name of the dataset to query}
}
\value{
a \code{data.frame}
}
\description{
Retrieve Data From SNCF API
}
\examples{
\dontrun{
get_sncf("ponctualite-mensuelle-transilien")
}
}
