% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_na.R
\name{impute_na}
\alias{impute_na}
\title{Remplace valeurs manquantes par Moyenne ou Médiane}
\usage{
impute_na(x, method = c("moyenne", "mediane"))
}
\arguments{
\item{x}{numeric vector}

\item{method}{choise between mean and median}
}
\value{
a numeric
}
\description{
Remplace valeurs manquantes par Moyenne ou Médiane
}
\examples{
impute_na(c(2, 3, 3, NA, 5))
}
